slackbot/requirements.txt
pandas
seaborn
requests
python-dotenv
sqlalchemy
sqlalchemy-utils
psycopg2-binary

